Refactors state module to pass refactored tests
Change code to ensure that test that unstarted machine raises the
MachineNotStarted exception and that test for transition passes.
A transition involves a source state and destination state.
Refactor involves mapping event, state -> transition for every substate
of the source state to ensure that the event will be handled by its substates
as well.