Rewrote tree module
The tree class is essentially a dictionary holding vertex
value objects. The vertex object is identified by its unique
id and holds ids of its children and parent. It optionally
holds data. We use the tree class to store state objects
and calculate the execution paths for the state machine
when transition happens.