Implements logic to add transitions to state machine
Transitions are now a callable object. They are connected
to the triggering signal. Transitions are called when the
source state "sends" the signal to the transition (receiver).
As signals are the only refenrences to transitions, we have
to ensure that we set the weakref parameter to False in
the signal.connect method.