Refactors repository and logic on Entity class
The StateRepository.insert method now takes in entity_name
directly as a non-keyword only arg. To specify a node in the
state tree, we also pass in name and state as keyword args
to the method.

Similarly, the .get method takes in entity_name as a non-keyword
arg only with name as a keyword arg to specify which vertex in the
tree to retrieve. After retreiving the vertex, it returns the
Vertex.data field.