Changed signature on set_head method
The change in signature is to accomodate it as a receiver
for HANDLED signal and also as a method which can be called
directly. We overload the sender parameter so that if its
argument is a Transition object, it looks at the second argument
name to indicate which vertex it is pointing to.